pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crm-medallion"
version = "0.1.0"
description = "CRM Data Medallion Framework - Multi-layered data processing pipeline for CRM data cleaning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "CRM Medallion Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "chardet>=5.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
llm = [
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
]
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
all = [
    "crm-medallion[llm,rag]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
crm-medallion = "crm_medallion.cli.main:cli"

[project.urls]
Homepage = "https://github.com/crm-medallion/crm-medallion"
Documentation = "https://github.com/crm-medallion/crm-medallion#readme"
Repository = "https://github.com/crm-medallion/crm-medallion"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/crm_medallion"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["crm_medallion"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
